kaos_py/pyproject.toml
[tool.poetry]
name = "kaos-py"
version = "0.1.0"
description = ""
authors = ["Gökmen Görgen <gkmngrgn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.0.1"
pillow = "^10.4.0"

[tool.poetry.scripts]
kaos = "kaos_py:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
